Add a "clear" button and related code for the OutputView
Summary:
Adds a clear button and related code for the OutputView,
sometimes the amount of text inside of the OutputView is just
too much and we wanna clear, not filter, the dialog to be
able to look at new strings comming in.

Also, a few cleanups on some files regarding dead code and
micro-opts.

Signed-off-by: Tomaz Canabrava <tcanabrava@kde.org>

Test Plan: manual testing

Reviewers: kfunk, apol

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4126