File templates assistant: focus first edit field on switch to new page
Summary:
Most pages have fields which are expected to be edited, so setting the
keyboard focus to the first item on the page saves the user some work
on average. With the Enter key being bound to the "Next" button, this
should also not add work to pages where default settings are okay.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, arrowdodger, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4287