Remove BaseType typedef from *Type classes
The documentation on AbstractType said subclasses must have a BaseType
typedef. However, very few types actually had it, and nothing used it
anymore (apparently only the oldcpp code used it in the past).

So I'm removing the remaining BaseType typedefs, and adjusting the doxygen
comment that says you need it.

Closes: D4274