Fix build when using AbstractUseBuilder::newUse(NameT*)
AbstractUseBuilder has multiple overloads for the newUse method:
- newUse(T*, RangeInRevision&, DeclarationPointer&), which does the real
work, and most language plugins use directly
- newUse(T*, DeclarationPointer&), which calls the 3-arg one, and is used
by the php plugin
- newUse(NameT*), which calls the 3-arg one, and for which I couldn't find
any user.

The one-argument overload didn't compile, but since this is a template,
the particular error it has only appears if the method is ever called; and
since none of the current language plugins calls it, nobody had noticed
before.

With this fix, newUse(NameT*) compiles and even works on a real plugin,
creating uses correctly.