Template file dialog: kill old widgets on re-entering TemplateOptionsPage
Summary:
Simple temporary fix for the current logic as triggered when going "Back"
from the options page to the previous and then entering again, which will
result in another call of the TemplateOptionsPage::load() method.
This should see some redesign possibly, but for this at least avoids
old layout & widget instances creating troubles.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4413