Allow to configure import commit message for all DVCS systems
Summary:
Currently the standard vcs import metadata widget for DVCS
has not supported the editing of the message used for the import
commit, other than the custom widget for SVN or CVS.
Instead the message was hardcoded.

This patch changes this to allow editing of the commit message
also for any DVCS using DvcsImportMetadataWidget, by
extending it with a text edit field for the message
and changing the code doing the commit to use the existing
message query API to get the commit message.

For having some default text in the editor, also for SVN & CVS
which currently start with an empty message, the
VcsImportMetadataWidget API gets also a setter for the message,
by which then a global default message is preset in the editor.

Reviewers: #kdevelop, brauch

Reviewed By: #kdevelop, brauch

Subscribers: brauch, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4513