Use consistently the same icon for debug launch everywhere
Summary:
"debug-run" is used in the toolbar and main menu for the Debug launch
action ~~and also for the Debug area~~ (actually the icon for the Debug area is also the other/old one, I confused myself as I use this for some time already :) ).
Thus it should also be used the launch config dialog with the Debug subentry, to not confuse
but show this is about the same thing.

Reviewers: #kdevelop, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4583