GrepOutputView: show older search results fully expanded initially
Summary:
When switching to a previous search result or when switching back
to the latest, the initial presentation of the search result
is completely collapsed. Which both does not match the initial
presentation of a new search result and also might conflict
with the usual workflow, where people want to see the complete
result fully expanded.

Ideally the complete view state would be remembered, so one
can quickly switch between search results to compare things,
but that needs more work. For now this simple fix should improve
things for most (extrapolated personal experience ;) ).

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D4605