Improve code completion of paths with dashes in their name
The model simply aborted code completion too soon, as it thought
that the dash stops a word. But when we do include-path completion
we actually want to not do that.

This patch implements that by reusing our existing include path
parser and then overwrites the corresponding model functions.

Note that the code completion box still "hides" after writing the
first "-", which I have not yet figured out why. Manual code
completion works in this case now, though. In general, this patch
improves quite a few things already so I'm pushing this early.

CCBUG: 376461