Pressing "Continue" starts "Debug" if program is not running
Changes the kdevdebuggershellui to make the "continue" action visible when the debugger is not running, otherwise the "continue" action would be disabled.
Triggering the "continue" action when while no debug is running will have the same effect as triggering "debug" action. The "continue" action for running debugs remains the same as before.

BUGS: 342245
Differential Revision: https://phabricator.kde.org/D4334