Add all .pro files to qmake project root file item.
This was done already for subfolders, but not for the project root.
This greatly improves the quality of the include/defines lookups
for projects that contain multiple .pro files in the project root.

Before, we simply picked the first .pro, which may not be what the
user actually selected in the "Open Project" dialog. Also, we may
encounter multiple .pro files, e.g. separate ones for unit tests,
in the project root directory.

Now, we check all .pro files for include paths and defines.