Use {} as default value for zero-set flags, not nullptr
Summary:
For some human code readers nullptr seems akward with bitflags,
(even more as QFlags is int-based -> sizeof). On quick reading
seeing a nullptr with unknown type even might result in wrong
assumption of a pointer type.
So for mental sanity of some of us using default constructor instead,
which still needs people to know it defaults to no bits set,
but that might be expected, as 0 is a usual default.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5161