Drop Qt4 app templates, add Qt5/QMake QWidgets app template
Summary:
Qt4 is no longer maintained, so providing app templates based
on it in the default KDevelop installation will not really
serve a lot of people.
Worse, due to alphabetical sorting the Qt4 variants are the
default selected for the categories Qt/Graphical, giving
a bad first impression.

Test Plan: Created and ran a new app with the new template.

Reviewers: #kdevelop, apol

Reviewed By: #kdevelop, apol

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5276