Drop explicit %{PROJECTDIR}/ from templates' ShowFilesAfterGeneration
Summary:
Using the var with the ShowFilesAfterGeneration entry is not needed
starting with 5.1.1 (again), as it will be automatically prepended
for all relative paths, as defined in kdevplatform/plugins/appwizard/README.dox

Not using the var makes files easier to read & avoids typos
in the var name. So let default templates use recommended notation.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5531