Add openDeclaration(Identifier) overload to AbstractDeclarationBuilder
As discussed on the mailing list, openDeclaration(QualifiedIdentifier)
makes little sense semantically, and in practice it's just taking the last
Identifier from the QID and using that.

Make it take an Identifier instead, leaving a QualifiedIdentifier overload
calling openDeclaration(id.last()) for compatibility. Also, print a warning
if it's called with a QualifiedIdentifier that actually has more than one
Identifier in it (I haven't seen this happen).