Create KAboutData object only after QApp instance, for working translations
Summary:
i18n() calls gettext, which only after some setlocale(LC_ALL, ""); starts
to look for env variables when deciding on the language catalog to use.
That call is done during QApp instance creation.

See also https://phabricator.kde.org/D5439#102005 and
https://phabricator.kde.org/D5455

Discussion if that is supposed to be like that has to be continued,
but for now this change should keep KDevelop on the safe side.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kfunk, ltoscano, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5598