Remove Q_SCRIPTABLE tags not used for D-Bus or QML (no more Kross support)
Summary:
There were lots of Q_SCRIPTABLE tags which once were added for Kross,
but no longer serve any purpose, besides a few being picked up/reused
for interfaces exposed to QML or D-Bus.

This patch removes any such tags which were not used or useful (due
to incomplete types) from QML and D-Bus.
For access to KDevelop::ICore's projectController from QML instead
a QObject property is now noted. And the methods of IProjectController
used from QML are tagged as Q_INVOKABLE, so they are not included
in the D-Bus interface.

It will also remove the console warnings during start-up
"
Invalid return type in method "core"
"
which are from method generateInterfaceXml(...) in
qtbase/src/dbus/qdbusxmlgenerator.cpp when it generates
the interface for ExternalScriptPlugin, which inherits
KDevelop::IPlugin, which had its method
ICore *core() const; tagged with Q_SCRIPTABLE,
and as the ExternalScriptPlugin registers itself as
object to the bus using QDBusConnection::ExportScriptableSlots
that method was included in the interface, but without ICore*
being known.

Reviewers: #kdevelop, apol

Reviewed By: #kdevelop, apol

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5604