Make it possible to open directories from the command line
Summary:
When using the terminal, it's sometimes useful to open the project we're in
with KDevelop. Running `kdevelop .` would do this in this case.
If kdevelop isn't running then it uses a fresh instance, if it does it tells
the ones we're using to open the project, just like for files.

Reviewers: #kdevelop, vhanda, mwolff

Reviewed By: #kdevelop, mwolff

Subscribers: mwolff, kfunk, cullmann, zhigalin, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5754