Make sure we don't get an empty element in the context menu
The outputwidget has a qwidgetaction, this action will be displayed both
in the toolbar and in the context menu. There's a limitation by QWidgets
that the widget will only be shown once.
Includes a name and an actual trigger action for the action that will be
shown in its defect. The text will be the widget placeholder and the
action will be to set the focus to the actual widget