Adds option to pass cursor line and column to external scripts
This patch adds a way to pass cursor line and column to external
scripts. The most straightforward use case is to open the file in vim
for making some kind of editing that is unavailable through vi-edit
mode. To this end, one would use the following command:

konsole -e vim -c "call cursor(%l, %c)" %f

It will open vim in konsole with vim's cursor in the same line and
column of the current document view.

Differential Revision: https://phabricator.kde.org/D5644