Fix CompilerFilterStrategy::errorInLine() to handle paths with whitespace
Summary:
The "([^: \\t]+):" does not match strings with white space. While this
reduces the chance of wrong matches, it also excludes valid ones for
paths with white space in them (like on new KDE CI).

The different regex for the different tools are a bit inconsistent on that
matter, so as consistent solution whitespaces in paths are supported.

Tests are adapted to test with both paths with spaces and paths without
(for now only unix-style, given the patterns currently used).

Test Plan:
test_filteringstrategy no longer fails when sources are below a path
with white spaces. And cmake/gcc errors in projects with such paths
are now also properly supported.

Reviewers: #kdevelop, volden, apol, kfunk

Reviewed By: #kdevelop, apol, kfunk

Subscribers: apol, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5995