Properly handle remote paths when creating nodejd module directories
When we open a remote path, we get one more segment for the host,
user and related information. As such, we always will have at least
two segments. Previously, we looped indefinitely and consumed a lot
of memory until that was exhausted and kdevelop crashed. Now we
handle this situation gracefully and users can again edit JS files
via fish/sftp remotely.

BUG: 369573
FIXED-IN: 5.1.2