Extend ClangClassHelper to offer both operator== & operator!= operators
Summary:
I do not know a best practice rule which recommends this.
But when I add some operator== implementation, I also add an
operator!= implementation to a class.
Because if users of the class API need the first, chance is
other users might want the inverse for other logic.

So I would find it handy to have codegen allow me to generate both
signatures.

Reviewers: #kdevelop, kfunk

Reviewed By: #kdevelop, kfunk

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D6346