Fix crash on clicking "Show documentation for xyz" in code tooltip
Summary:
Sync handling of focus change on opening the documentation docker
results in destruction of AbstractNavigationWidget object
and thus also the AbstractNavigationContext object it holds,
which is a QSharedData subclass, this way crashing the
QExplicitlySharedDataPointer on itself returned from the method
once it gets to its destructor.

Test Plan:
No longer crashes when documentation docker is closed and the
"Show documentation for xyz" link is activated.

Reviewers: #kdevelop, apol

Reviewed By: #kdevelop, apol

Subscribers: apol, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D6436