Fix splitting of root installation command, add default value for the same in .kcfg
Summary:
-KShell::splitArgs is used instead of string.split( )
Reason:
When suCommand is empty, the flow of control does not reach body of  "if(suCommandWithArg.isEmpty() )"

If QString is empty and it is split using .split() method, the returned QStringList is non empty,
that is not the case with KShell::splitArgs.

We could use if( suCommand.isEmpty() ) instead but since we ultimately have to return suCommandWithArg,
we must have a check for emptiness of suCommandWithArg, and using KShell::splitArgs eliminates the need for having
to check if( suCommand.isEmpty() ).

-Added a default value for suCommand in makebuilderconfig.kcfg so that when new projects are
created, the entry suCommand isn't empty.

Test Plan:
-Without the patch when suCommand entry is empty, "" is used as root installation command when project is installed as root (hence installation fails)
-With the patch, kdesu -t is used.

Reviewers: apol

Reviewed By: apol

Subscribers: kfunk, apol, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D5777