WelcomePage: Don't load QML in some scenarios
There are a couple of distros which still provide x86 binaries with SSE2
support disabled to support older architectures. Let's not render
KDevelop useless on these platforms, this is easy enough to play along
with.

If one attempts to load QML on a processor without SSE2 support, the Qt
V8 engine by defaults just aborts the program.

Mitigate this by just not loading any QML in case the processor has no
SSE2 support.

BUG: 381999
FIXED-IN: 5.1.2
Ref T6462