Fix duplicated annotation context menus & entries on repeated border show
Summary:
Handler of annotationContextMenuAboutToShow was not disconnected,
thus accumulating connect calls and that way repeated addition
of on-the-fly-created actions "Copy Revision" & "History...".

Test Plan:
Before on multiple hide and show of the annotation border for the same
document the two actions mentioned above were duplicated by times of
shown border. With this patch they no longer are. No sideeffects seen.

Reviewers: #kdevelop, brauch

Reviewed By: #kdevelop, brauch

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D6838