clang: fix precompiled preamble cache misses
We were a) passing in a wrong file size, and b) a different set of unsaved
files on building the translation unit as compared to invoking completion.
This resulted in the precompiled preamble being rebuilt every single time.
On my test file, time spent in clang_codeCompleteAt goes down from ~700ms
to ~12ms with this change after the project is fully parsed.

It still doesn't work in header files at all.

Differential Revision: https://phabricator.kde.org/D6905