Restore preview for source formatting styles in config selection
While 4a48b37214e088ee9aacda5b8798b927df6d290b in 2011 used the new
usePreview flag of SourceFormatterStyle to disable the preview,
and only for kdev_format_sources.sh, actually the default value of the
flag has been set already to false in the matching commit to kdevplatform,
at least to the (unused) default constructor.
And at least once the SourceFormatterStyle(const QString &name) constructor
was fixed to not leave the initial value of the usePreview flag undefined
and now sets it to false per default as well, all the other styles
lost their previews as well.

Explicitely setting the flag to true for those styles where a preview
is usable restores the preview.