Fix for Outoutview CmakeMultiline expression accidentally hits on CMake Error.
Summary:
The expression intended for catching CMakeMultiline errors e.g.:

  CMake Error: Error in cmake code at
  /some/path with spaces/to/a/project/CMakeLists.txt:10:

Where the regex would hit on the second line.

Now it also hits on lines such as this:

CMake Error at SecondSubDir/CMakeLists.txt:1:

Which means that the filename is extracted wrong and the user will be sent to a new file instead of the file with the error.

Changed the regex in question and added a unittest that catches this scenario.

Reviewed By: kfunk

Subscribers: volden, kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D7554