Enable reproducible packaging of template archives
Summary:
Builds of KDevelop are not reproducible right now because the archives
containing templates contain metadata related to the build environment,
such as the time of checkout. The order of files in an archive is also
not guaranteed.

The Reproducible Builds project recommends [1] to use (in addition to
what we have) --mtime="@${SOURCE_DATE_EPOCH}" and --sort=name. This will
only be used if the environment variable is set. The latter option is
available only since tar 1.28 (from 2014), but we assume that people
have it if they know about reproducible builds.

[1] https://reproducible-builds.org/docs/archives/

Reviewers: #kdevelop, kfunk

Subscribers: kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D7693