OutputFilteringStrategies: make parsing of cmake output for project directory work on Windows
Summary:
This is a follow-up to https://phabricator.kde.org/D7040 and adds support for parsing the cmake project directory on Windows.
Also relates to https://bugs.kde.org/show_bug.cgi?id=321982.

Changes:

extended regex to filter output as produced on Windows systems
cmake binary has extension .exe or .bat (in craft environment)
project paths start with [A-Za-z]: instead of /
added creation of different styles of Windows paths to testlinebuilderfunctions.h
modified TestFilteringStrategy::testCompilerFilterstrategyUrlFromAction_data() to also test with Windows paths
NOTE: it doesn't make sense to test Windows paths on Linux and vice versa, as class Path, that is used to store paths, is not platform agnostic and handles paths differently depending on the system it is compiled for
instead system specific paths are used during testing, depending on the platform the tests are compiled for

Reviewed By: volden

Subscribers: volden, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D7615