Prefer the first available compiler as default compiler
Commit 65106becddd introduced this regression: The code used to
return early when a compiler was found. The new loop to cache
the result always iterates over all compilers.

Essentially, this reverts the whole logic within the IDM: Instead
of prefering, say, Clang over GCC, it actually does the opposite
and will prefer GCC over Clang. This patch adds the missing break
to restore the old behavior.

This ensures we will once again use Clang instead of GCC by default
for includes and built-in defines. As exemplified by the recent
breakage (__has_include) when using GCC, and the history of the
shaky GCC integration, this is important for a good default
impression of KDevelop.