Ensure setting of custom app language is seen on start-up
The custom language setting is read in from the config in a method hooked
up to the QApp instance creation (see KXMLGUI's kswitchlanguagedialog_p.cpp
sources), which deploys the custom setting by setting the LANGUAGE env var.

For this to work, the respective init code of KLocalizedString static
data, which contains the set of languages as read from the env vars,
needs to be only triggered after the QApp instance creation.

Which includes calling KLocalizedString::setApplicationDomain() only then,
which this patch cares for.