Add missing X-KDevelop-Category entries to plugins' metadata
Summary:
This entry is defined as required (cmp. IPlugin API docs) with values either
"Global" or "Project". Having it not set and thus being treated with a value
of empty string results e.g. in plugins not being loaded or unloaded after
the user enabled or disabled plugins in the settings.

Reviewers: #kdevelop, brauch

Reviewed By: #kdevelop, brauch

Subscribers: brauch, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D8139